showmehow: Connect to showmehow-as-a-service
Note that in this changewe we need to be more rigorous with
NONINTERACTIVE. In that case, we can't even re-connect to the
because we will be in the same process that started the service,
which is also blocking on our response.

The only work that this commandline utility does now is
to interact with the service to deliver the lessons. It can be
used as a model for other parts of the system that want to
integrate with the service.