showmehow: Re-implement as an async state machine
Previously, showmehow was entirely synchronous and used
blocking input as a means to control lesson flow. This doesn't
really work when we start adding signals about external OS events
into the mix.

Things were already bad with the out-of-band ServiceMonitor class
and mixing GDBus and threads doesn't work very well.

In reality, going through the showmehow narrative arcs is really
a state machine and we can model it that way. This allows us
to say that external events might be inputs to the state machine
which cause it to do things if it is in the waiting state.