showmehow: Import readline and use input() for input
Previously we did this on the event loop when we had to care about
external events, but now since that has been moved out into c-g-s
we can just block and use input() again. Doing so has the benefit
of allowing us to import readline, which overrides the implementation
of input() to allow for history and line editing.

In the future we could also allow tab completion. We enable that
for now with filenames.