showmehow: Require the caller to move to the next task after _initialize
_initialize could be called in a few different contexts and is
really about setting up sessions. We don't want it to start the
next task straight away in the case that we're just about to
call start(), but we also don't want it to only move on to the
next task in the case that a session was set previously - some lessons
don't have sessions.

https://phabricator.endlessm.com/T16207